added a generic Acoustic chunk class to remove redundances between acoustic file and acoustic chunk
