+++ pypam/__init__.py
@@ -1,5 +1,6 @@
 from pypam.dataset import DataSet
 from pypam.acoustic_survey import ASA
 from pypam.acoustic_file import AcuFile
+from pypam.acoustic_chunk import AcuChunk, GenericAcuFile
 from pypam.signal import Signal
 from pypam.detection import Detection
